sightseeing_guide_agent/requirements.txt
google-adk==1.14.0
google-cloud-logging
python-dotenv
langchain
langchain-community
wikipedia